First attempt at using gene_trees.R together with TreeShrink
Refactor rules so that TreeShrink and gene_trees.R (and subsequent
processing) have separate rules. This necessary as TreeShrink only
considers terminal branches when identifying 'outliers' whereas there
an be also paralogous sequences in an MSA. gene_trees.R script
therefore checks only the internal branches after a TreeShrink run to
identify long internal branches that may have resulted from paralogs
being included in the alignment.

Also implement a sequence elimination step after running two 'outlier
detection' rules for the sequences that have 95% or more gaps. This
can happen after running TrimAl if 'fragmented' sequences are used as
input to the pipeline and will cause problems since sequences that are
that short compared to the rest of the MSA can and often will be
aligned spuriously.